requirements.txt
pymupdf
paddleocr
Pillow
langchain_groq
langchain_community
langchain
python-dotenv
tqdm
streamlit
pandas
tabulate
pdf2image==0.1.4
paddlepaddle
sentence-transformers
chromadb
protobuf==3.20.3
torch==2.1.0
pysqlite3-binary